fix: Move TableFormat into spicebench (#61)
* fix: Move TableFormat into spicebench

* chore: auto-fix cargo fmt + clippy

---------
